+++ src/EV/ev.h
@@ -1,7 +1,7 @@
 // ev.h
 #ifndef EV_H
 #define EV_H
-#include "vmu.h"
+#include "../VMU/vmu.h"
 
 void handle_signal(int sig);
